MG-MINEGAME-V136
- Update

+++ src/me/nao/manager/mg/GameIntoMap.java
@@ -586,7 +586,7 @@ public class GameIntoMap {
 		int puntos2 = gp.getDeads();
 		int puntos3 = gp.getRevive();
 		int puntos4 = gp.getHelpRevive();
-		int puntos5 = gp.getDamage();
+		long puntos5 = gp.getDamage();
 		player.sendMessage(""+ChatColor.YELLOW+ChatColor.BOLD+" [PUNTOS DE LA PARTIDA]");
 		player.sendMessage(""+ChatColor.GREEN+" Eliminaciones : "+ChatColor.YELLOW+puntos);
 		player.sendMessage(""+ChatColor.GREEN+" Muertes : "+ChatColor.YELLOW+puntos2);
